ClimaComms

ClimaComms.ClimaComms — Module

ClimaComms

Abstracts the communications interface for the various CliMA components in order to:

enable the use of different backends as transports (MPI, SharedArrays, etc.), and

transparently support single or double buffering for GPUs, depending on whether the transport has the ability to access GPU memory.

Use one of the ClimaComms backends, currently:

ClimaCommsMPI uses MPI-2 asynchronous primitives through MPI.jl and supports a single buffer if the MPI implementation is CUDA-aware.

ClimaCommsSA uses Julia's Distributed capability and SharedArrays.jl for within-node communication.

Contexts

ClimaComms.AbstractCommsContext — Type

AbstractCommsContext

The base type for a communications context. Each backend defines a concrete subtype of this, an instance of which is constructed by passing in an array of Neighbors.

Communication interface

ClimaComms.init — Function

init(::Type{CC}) where {CC <: AbstractCommsContext}

Perform any necessary initialization for the specified backend. Return a tuple of the processor ID and the number of participating processors.

ClimaComms.mypid — Function

mypid(::Type{CC}) where {CC <: AbstractCommsContext}

Return the processor ID.

ClimaComms.iamroot — Function

iamroot(::Type{CC}) where {CC <: AbstractCommsContext}

Return true if the calling processor is the root processor.

ClimaComms.nprocs — Function

nprocs(::Type{CC}) where {CC <: AbstractCommsContext}

Return the number of participating processors.

ClimaComms.singlebuffered — Function

singlebuffered(::Type{CC}) where {CC <: AbstractCommsContext}

Returns true if communication can be single-buffered.

ClimaComms.start — Function

start(ctx::CC; kwargs...) where {CC <: AbstractCommsContext}

Initiate communication. The stage areas of all the send RelayBuffers must be filled before this is called!

ClimaComms.progress — Function

progress(ctx::CC) where {CC <: AbstractCommsContext}

Drive communication. Call after start() to ensure that communication proceeds asynchronously.

ClimaComms.finish — Function

finish(ctx::CC; kwargs...) where {CC <: AbstractCommsContext}

Complete the communications step begun by start(). After this returns, data received from all neighbors will be available in the stage areas of each neighbor's receive buffer.

ClimaComms.barrier — Function

barrier(ctx::CC) where {CC <: AbstractCommsContext}

Perform a global synchronization across all participating processors.

ClimaComms.reduce — Function

reduce(ctx::CC, val, op) where {CC <: AbstractCommsContext}

Perform a reduction across all participating processors, using op as the reduction operator and val as this rank's reduction value. Return the result to the first processor only.

ClimaComms.abort — Function

abort(ctx::CC, status::Int) where {CC <: AbstractCommsContext}

Terminate the caller and all participating processors with the specified status.

Neighbors

ClimaComms.Neighbor — Type

Neighbor

Represents a communications neighbor. Constructors are defined by the backends and require a unique ID that represents the neighbor, the array type to be used for staging (usually Array on the CPU and CuArray for GPU code), the element type for the array, and the dimensions for the send and receive buffers.

Missing docstring.

Missing docstring for ClimaComms.id. Check Documenter's build log for details.

ClimaComms.send_stage — Function

send_stage(neighbor::Neighbor)

Return the staging area for data that is to be sent to the specified neighbor.

ClimaComms.recv_stage — Function

recv_stage(neighbor::Neighbor)

Return the staging area for data that has been received from the specified neighbor.