ClimaComms

ClimaComms.ClimaComms — Module

ClimaComms

Abstracts the communications interface for the various CliMA components in order to:

enable the use of different backends as transports (MPI, SharedArrays, etc.), and

transparently support single or double buffering for GPUs, depending on whether the transport has the ability to access GPU memory.

Use one of the ClimaComms backends, currently:

ClimaCommsMPI uses MPI-2 asynchronous primitives through MPI.jl and supports a single buffer if the MPI implementation is CUDA-aware.

ClimaCommsSA uses Julia's Distributed capability and SharedArrays.jl for within-node communication.

Contexts

ClimaComms.AbstractCommsContext — Type

AbstractCommsContext

The base type for a communications context. Each backend defines a concrete subtype of this.

ClimaComms.AbstractGraphContext — Type

AbstractGraphContext

A context for communicating between processes in a graph.

Communication interface

ClimaComms.init — Function

(pid, nprocs) = init(ctx::AbstractCommsContext)

Perform any necessary initialization for the specified backend. Return a tuple of the processor ID and the number of participating processors.

ClimaComms.mypid — Function

mypid(ctx::AbstractCommsContext)

Return the processor ID.

ClimaComms.iamroot — Function

iamroot(ctx::AbstractCommsContext)

Return true if the calling processor is the root processor.

ClimaComms.nprocs — Function

nprocs(ctx::AbstractCommsContext)

Return the number of participating processors.

ClimaComms.barrier — Function

barrier(ctx::CC) where {CC <: AbstractCommsContext}

Perform a global synchronization across all participating processors.

ClimaComms.reduce — Function

reduce(ctx::CC, val, op) where {CC <: AbstractCommsContext}

Perform a reduction across all participating processors, using op as the reduction operator and val as this rank's reduction value. Return the result to the first processor only.

ClimaComms.abort — Function

abort(ctx::CC, status::Int) where {CC <: AbstractCommsContext}

Terminate the caller and all participating processors with the specified status.

ClimaComms.graph_context — Function

graph_context(context::AbstractCommsContext, sendarray, sendpids, sendlengths, recvarray, recvpids, recvlengths)

Construct a communication context for exchanging neighbor data via a graph.

Arguments:

context: the communication context on which to construct the graph context.

sendarray: array containing data to send

sendpids: list of processor IDs to send

sendlengths: list of lengths of data to send to each process ID

recvarray: array to receive data into

recvpids: list of processor IDs to receive from

recvlengths: list of lengths of data to receive from each process ID

This should return an AbstractGraphContext object.

ClimaComms.start — Function

start(ctx::AbstractGraphContext)

Initiate graph data exchange.

ClimaComms.progress — Function

progress(ctx::AbstractGraphContext)

Drive communication. Call after start to ensure that communication proceeds asynchronously.

ClimaComms.finish — Function

finish(ctx::AbstractGraphContext)

Complete the communications step begun by start(). After this returns, data received from all neighbors will be available in the stage areas of each neighbor's receive buffer.

Contexts

ClimaComms.SingletonCommsContext — Type

SingletonCommsContext()

A singleton communications context, used for single-process runs.