APIs

ClimaComms.ClimaComms — Module

ClimaComms

Abstracts the communications interface for the various CliMA components in order to:

support different computational backends (CPUs, GPUs)

enable the use of different backends as transports (MPI, SharedArrays, etc.), and

transparently support single or double buffering for GPUs, depending on whether the transport has the ability to access GPU memory.

Loading

ClimaComms.@import_required_backends — Macro

ClimaComms.@import_required_backends

If the desired context is MPI (as determined by ClimaComms.context()), try loading MPI.jl. If the desired device is CUDA (as determined by ClimaComms.device()), try loading CUDA.jl.

ClimaComms.cuda_is_required — Function

cuda_is_required()

Returns a Bool indicating if CUDA should be loaded, based on the ENV["CLIMACOMMS_DEVICE"]. See ClimaComms.device for more information.

cuda_is_required() && using CUDA

ClimaComms.mpi_is_required — Function

mpi_is_required()

Returns a Bool indicating if MPI should be loaded, based on the ENV["CLIMACOMMS_CONTEXT"]. See ClimaComms.context for more information.

mpi_is_required() && using MPI

Devices

ClimaComms.AbstractDevice — Type

AbstractDevice

The base type for a device.

ClimaComms.AbstractCPUDevice — Type

AbstractCPUDevice()

Abstract device type for single-threaded and multi-threaded CPU runs.

ClimaComms.CPUSingleThreaded — Type

CPUSingleThreaded()

Use the CPU with single thread.

ClimaComms.CPUMultiThreaded — Type

CPUMultiThreaded()

Use the CPU with multiple thread.

ClimaComms.CUDADevice — Type

CUDADevice()

Use NVIDIA GPU accelarator

ClimaComms.device — Function

ClimaComms.device()

Determine the device to use depending on the CLIMACOMMS_DEVICE environment variable.

Allowed values:

CPU, single-threaded or multi-threaded depending on the number of threads;

CPUSingleThreaded,

CPUMultiThreaded,

CUDA.

The default is CPU.

ClimaComms.device_functional — Function

ClimaComms.device_functional(device)

Return true when the device is correctly set up.

ClimaComms.array_type — Function

ClimaComms.array_type(::AbstractDevice)

The base array type used by the specified device (currently Array or CuArray).

ClimaComms.free_memory — Function

ClimaComms.free_memory(device)

Bytes of memory that are currently available for allocation on the device.

ClimaComms.total_memory — Function

ClimaComms.total_memory(device)

Bytes of memory that are theoretically available for allocation on the device.

ClimaComms.allowscalar — Function

allowscalar(f, ::AbstractDevice, args...; kwargs...)

Device-flexible version of CUDA.@allowscalar.

Lowers to

f(args...)

for CPU devices and

CUDA.@allowscalar f(args...)

for CUDA devices.

This is usefully written with closures via

allowscalar(device) do
    f()
end

ClimaComms.@time — Macro

@time device expr

Device-flexible @time.

Lowers to

@time expr

for CPU devices and

CUDA.@time expr

for CUDA devices.

ClimaComms.@elapsed — Macro

@elapsed device expr

Device-flexible @elapsed.

Lowers to

@elapsed expr

for CPU devices and

CUDA.@elapsed expr

for CUDA devices.

ClimaComms.@assert — Macro

@assert device cond [text]

Device-flexible @assert.

Lowers to

@assert cond [text]

for CPU devices and

CUDA.@cuassert cond [text]

for CUDA devices.

ClimaComms.@sync — Macro

@sync device expr

Device-flexible @sync.

Lowers to

@sync expr

for CPU devices and

CUDA.@sync expr

for CUDA devices.

An example use-case of this might be:

BenchmarkTools.@benchmark begin
    if ClimaComms.device() isa ClimaComms.CUDADevice
        CUDA.@sync begin
            launch_cuda_kernels_or_spawn_tasks!(...)
        end
    elseif ClimaComms.device() isa ClimaComms.CPUMultiThreading
        Base.@sync begin
            launch_cuda_kernels_or_spawn_tasks!(...)
        end
    end
end

If the CPU version of the above example does not leverage spawned tasks (which require using Base.sync or Threads.wait to synchronize), then you may want to simply use @cuda_sync.

ClimaComms.@cuda_sync — Macro

@cuda_sync device expr

Device-flexible CUDA.@sync.

Lowers to

expr

for CPU devices and

CUDA.@sync expr

for CUDA devices.

Adapt.adapt_structure — Method

Adapt.adapt_structure(::Type{<:AbstractArray}, device::AbstractDevice)

Adapt a given device to a device associated with the given array type.

Example

Adapt.adapt_structure(Array, ClimaComms.CUDADevice()) -> ClimaComms.CPUSingleThreaded()

Note

By default, adapting to Array creates a CPUSingleThreaded device, and there is currently no way to conver to a CPUMultiThreaded device.

Threading

ClimaComms.@threaded — Macro

@threaded [device] [kwargs...] for ... end

Device-flexible generalization of Threads.@threads, which distributes the iterations of a for-loop across multiple threads, with the option to control thread coarsening and GPU kernel configuration. Coarsening makes each thread evaluate more than one iteration of the loop, which can improve performance by reducing the runtime overhead of launching additional threads (though too much coarsening worsens performance because it reduces parallelization). The device is either inferred by calling ClimaComms.device(), or it can be specified manually, with the following device-dependent behavior:

When device is a CPUSingleThreaded(), the loop is evaluated as-is. This avoids the runtime overhead of calling Threads.@threads with a single thread, and, when the device type is statically inferrable, it also avoids compilation overhead.

When device is a CPUMultiThreaded(), the loop is passed to Threads.@threads. This supports three different kinds of "schedulers" for determining how many iterations of the loop to evaluate in each thread:

(default) a "dynamic" scheduler that changes the number of iterations as new threads are launched,

a "static" scheduler that evaluates a fixed number of iterations per thread, and

a "greedy" scheduler that uses a small number of threads, continuously evaluating iterations in each thread until the loop is completed (only available as of Julia 1.11).

Setting coarsen to :dynamic or :greedy launches threads with those schedulers. Setting it to nothing or an integer value launches threads with static scheduling (using nothing is similar to using 1, but slightly more performant). To read more about multi-threading, see the documentation for Threads.@threads.

When device is a CUDADevice(), the loop is compiled with CUDA.@cuda and run with CUDA.@sync. Since CUDA launches all threads at the same time, only static scheduling can be used. Setting coarsen to any symbol causes each thread to evaluate a single iteration (default), and setting it to an integer value causes each thread to evaluate that number of iterations (the default is similar to using 1, but slightly more performant). If the total number of iterations in the loop is extremely large, the specified coarsening may require more threads than can be simultaneously launched on the GPU, in which case the amount of coarsening is automatically increased.

The optional argument block_size is also available for manually specifying the size of each block on a GPU. The default value of :auto sets the number of threads in each block to the largest possible value that permits a high GPU "occupancy" (the number of active thread warps in each multiprocessor executing the kernel). An integer can be used instead of :auto to override this default value. If the specified value exceeds the total number of threads, it is automatically decreased to avoid idle threads.

Any iterator with methods for length and getindex can be used in a @threaded loop, as long as getindex supports one-based linear indexing. All lazy iterators from Base and Base.Iterators, such as zip, enumerate, Iterators.product, and generator expressions, are also compatible with @threaded. (Although these iterators do not define methods for getindex, they are automatically modified by threadable to support getindex.) Using multiple iterators with @threaded is equivalent to looping over a single Iterators.product, with the innermost iterator of the loop appearing first in the product, and the outermost iterator appearing last.

NOTE: When a value in the body of the loop has a type that cannot be inferred by the compiler, an InvalidIRError will be thrown during compilation for a CUDADevice(). In particular, global variables are not inferrable, so @threaded must be wrapped in a function whenever it is used in the REPL:

julia> a = CUDA.CuArray{Int}(undef, 100); b = similar(a);

julia> threaded_copyto!(a, b) = ClimaComms.@threaded for i in axes(a, 1)
           a[i] = b[i]
       end
threaded_copyto! (generic function with 1 method)

julia> threaded_copyto!(a, b)

julia> ClimaComms.@threaded for i in axes(a, 1)
           a[i] = b[i]
       end
ERROR: InvalidIRError: ...

Moreover, type variables are not inferrable across function boundaries, so types used in a threaded loop cannot be precomputed before the loop:

julia> threaded_add_epsilon!(a) = ClimaComms.@threaded for i in axes(a, 1)
           FT = eltype(a)
           a[i] += eps(FT)
       end
threaded_add_epsilon! (generic function with 1 method)

julia> threaded_add_epsilon!(a)

julia> function threaded_add_epsilon!(a)
           FT = eltype(a)
           ClimaComms.@threaded for i in axes(a, 1)
               a[i] += eps(FT)
           end
       end
threaded_add_epsilon! (generic function with 1 method)

julia> threaded_add_epsilon!(a)
ERROR: InvalidIRError: ...

To fix other kinds of inference issues on GPU devices, especially ones brought about by indexing into iterators with nonuniform element types, see UnrolledUtilities.jl.

ClimaComms.@shmem_threaded — Macro

@shmem_threaded ... = ...

Extension of @threaded for working with shared memory arrays, which can be used immediately before a @threaded loop in the following manner:

outer_iterator_1 = ...
outer_iterator_2 = ...
...
@threaded device ... begin
    @shmem_threaded inner_iterator = ...
    for outer_item_1 in outer_iterator_1, outer_item_2 in outer_iterator_2, ...
        ...
        array_1 = shmem_array(inner_iterator, ...)
        array_2 = shmem_array(inner_iterator, ...)
        ...
        for inner_item in inner_iterator
            <write to arrays based on inner and outer loop items>
        end
        ...
        for inner_item in inner_iterator
            <read from and write to arrays based on inner and outer loop items>
        end
        ...
        for inner_item in inner_iterator
            <read from arrays based on inner and outer loop items>
        end
        ...
    end
end

When an iterator variable assignment is annotated with @shmem_threaded, the iterator's elements are distributed across the threads in each block of a GPU, which have access to the same shared memory arrays. Iterators in the outer @threaded loop are then only distributed across GPU blocks, instead of being distributed across both blocks and threads within the blocks. On CPU devices, which do not provide blocks of threads with access to high-performance shared memory, the @shmem_threaded annotation does nothing.

An iterator whose assignment is annotated with @shmem_threaded can be used in loops nested inside the outer @threaded loop, with each GPU thread only processing a subset of the iterator's elements. Coarsening of the inner loops can be controlled by specifying the shmem_coarsen keyword argument for @threaded. Setting shmem_coarsen to :auto causes each thread to process one element of the iterator, and setting it to an integer value causes each thread to process that number of elements.

When the outer loop is not needed, it can be replaced by a begin ... end expression that is only distributed across the threads in a single GPU block:

@threaded device ... begin
    @shmem_threaded iterator = ...
    begin ... end
end

Since an iterator annotated with @shmem_threaded can be used to read data from shared memory arrays, every loop and reduction over the iterator is followed by a call to auto_sync! on GPU devices, ensuring that shared memory reads are consistent across threads.

When synchronization is not needed after a loop over an iterator annotated with @shmem_threaded, it can be disabled using disable_auto_sync. This can always be done for, e.g., the last such loop in a @threaded expression, since there are no subsequent loops that require consistency of shared memory. In general, disable_auto_sync should be called whenever doing so preserves correctness because unnecessary synchronization pauses can degrade performance.

Since each GPU thread processes a subset of the elements in an iterator annotated with @shmem_threaded, functions that are meant to process the entire iterator, like sum and minimum, would not generate valid results for its subsets. If functions like this need to be evaluated, their results should always be precomputed before @threaded loops.

When the size of an iterator annotated with @shmem_threaded is used to compute the dimensions of a shmem_array, it must be inferrable during compilation to avoid unnecessary runtime allocations. See StaticArrays.jl for how to construct statically-sized analogues of ranges and arrays.

ClimaComms.threaded — Function

threaded(f, device, itrs...; [shmem_threaded], kwargs...)

Functional form of @threaded. Depending on whether the @shmem_threaded iterator is specified, this is equivalent to one of the following expressions when n other iterators are provided:

@threaded device kwargs... begin
    for item_n in itrs[n], ..., item_2 in itrs[2], item_1 in itrs[1]
        f(item_1, item_2, ..., item_n)
    end
end

@threaded device kwargs... begin
    @shmem_threaded shmem_itr = shmem_threaded
    for item_n in itrs[n], ..., item_2 in itrs[2], item_1 in itrs[1]
        f(item_1, item_2, ..., item_n, shmem_itr)
    end
end

If only the shmem_threaded iterator is available, this is instead equivalent to the following expression:

@threaded device kwargs... begin
    @shmem_threaded shmem_itr = shmem_threaded
    begin
        f(shmem_itr)
    end
end

On single-threaded CPU devices, @threaded and @shmem_threaded annotations are eliminated without any intermediate function calls, so the expressions with macros will typically compile slightly faster than the threaded function. On other devices, the only differences between the macro expressions and the threaded function are syntactic:

the body of the @threaded loop is specified as a function, with return used in place of continue,

the @shmem_threaded iterator is specified as a keyword argument,

every symbol used as a keyword argument is wrapped in a Val, e.g., :auto is specified as Val(:auto).

Although the macro expressions are internally converted into threaded function calls, these differences make the functional form somewhat harder to read, so it should generally be avoided.

ClimaComms.threadable — Function

threadable(device, itrs...)

Modifies one or more iterators so that they each have the following properties:

a concrete element type

a well-defined length, and a corresponding method for isempty

support for linear indexing through getindex, with firstindex set to 1

When no specialized method of threadable is defined for some iterator type, an error will be thrown if the iterator does not already have these properties.

The threadable version of Iterators.product internally converts linear indices to Cartesian indices with Base.multiplicativeinverse on GPU devices, avoiding inefficient division operations. On CPU devices, this conversion is done using regular integer division.

ClimaComms.shareable — Function

shareable(itr, device)

Wrapper for a threadable iterator whose data can be shared across multiple threads on GPU devices. Most methods for iterators and AbstractArrays are forwarded to the underlying itr, but loops and reductions are modified so that each thread only interacts with a subset of the data in itr.

Specifically, when eachindex is called on a shareable iterator, the threads in each GPU block evenly divide the available indices among themselves. The corresponding methods for map, iterate (which is used by for-loops), and copyto! (which is used by broadcast expressions) all rely on eachindex to specify which items to loop over. Analogues of reduce! and mapreduce!, called shmem_reduce! and shmem_mapreduce!, can also be used to evaluate parallel reductions of shareable iterators on GPUs, with runtime that scales in proportion to the base-2 logarithm of iterator length.

Any broadcast expression over shareable iterators can have its output container automatically allocated in the form of a shmem_array. Non-shareable iterators can also be used in broadcast expressions with shareable ones, but the resulting output containers may not be inferred correctly if the non-shareable iterators have different dimensions than the shareable ones.

Every shareable iterator has an internal auto_sync flag that determines whether auto_sync! should be called after every loop or reduction. By default, this flag is always set to true, but in certain situations it can be set to false using disable_auto_sync, which may improve performance.

Since global memory is typically only used for storing the final outputs of computations, rather than caching intermediate data like shared memory, auto_sync! only synchronizes threads that modify shared memory arrays. If global memory arrays are used to cache intermediate data instead of shared memory arrays, loops and reductions may need to be explicitly synchronized using sync_shmem_threads!.

ClimaComms.set_metadata — Function

set_metadata(itr, metadata)

Adds metadata to a shareable iterator that can be inferred during compilation, allowing loops and reductions over the iterator to be unrolled.

ClimaComms.disable_auto_sync — Function

disable_auto_sync(itr)

Sets the auto_sync flag of a shareable iterator to false, so that the iterator can be used in a loop or reduction without being followed by a call to auto_sync!. This is safe to use in the following scenarios:

Looping over a shareable iterator without writing new data to shared memory or global memory that will need to be read by other threads. For example, a loop that just saves the final result of a computation to global memory should not be synchronized.

Running two consecutive loops that independently modify the same values in an array. For example, if a loop modifies the value at each index in an array, like array .= sin.(array), and then another loop modifies the value at each index again without reading data from adjacent indices, like array .*= 2, the first loop over the array should not be synchronized.

Applying a reduction operator to an array and then only using the result in @unique_shmem_thread expressions. Without synchronization, the result of any reduction operator is only guaranteed to be correct in threads that are part of the first warp in each GPU block. The implementation of @unique_shmem_thread guarantees that it will always choose one of these threads, so the expression it executes will have the correct result regardless of whether automatic synchronization is disabled. If the result is not needed in other threads, the reduction should not be synchronized.

ClimaComms.auto_sync! — Function

auto_sync!(itr)

Synchronizes GPU threads with access to the same shareable iterators, if the auto_sync flag of the given shareable iterator is set to true. On CPU devices, this function does nothing. Calls to auto_sync! are automatically inserted into most loops and reductions over shareable iterators, but custom loop functions may require explicit calls to auto_sync!.

ClimaComms.sync_shmem_threads! — Function

sync_shmem_threads!(device)

Internal function called by auto_sync! for shareable iterators whose auto_sync flags are set to true. Non-standard operations with loops or reductions may require explicit calls to sync_shmem_threads!.

ClimaComms.shmem_array — Function

shmem_array(itr, [T], [dims])
shmem_array(device, T, dims)

Device-flexible array whose element type T and dimensions dims are known during compilation, which corresponds to a static shared memory array on a GPU. Shared memory can be jointly modified by the threads in each GPU block, allowing kernels to have interdependencies between threads. On CPU devices, which do not provide access to high-performance shared memory, a shmem_array instead corresponds to a statically-sized MArray stored in register memory. Shared memory on a GPU and register memory on a CPU are both much faster to access and modify than the global memory used by Arrays and CuArrays.

The recommended method for this function uses a shareable iterator to determine the device, a default element type, and a default array size, and an alternative method that directly accepts those parameters is also available. However, the alternative method does not propagate additional metadata from the shareable iterator that is required for unrolling of loops and reductions, so using it disables automatic unrolling of operations on the resulting array. If the default element type or dimensions of a shareable iterator cannot be inferred during compilation, inferrable values of T or dims need to be passed to the shmem_array function to override the default values.

ClimaComms.@unique_shmem_thread — Macro

@unique_shmem_thread itr ...

Isolates an expression that should only be evaluated on a single thread in a @threaded loop with an @shmem_threaded iterator. For example, the result of a reduction over shared memory arrays can only be written to some specific address in global memory within a @unique_shmem_thread expression, since multiple threads simultaneously writing to the same index can lead to race conditions.

Any call to auto_sync! in a @unique_shmem_thread expression can lead to deadlock, as the thread that issues a request for synchronization will be stuck waiting for other threads that never issue one. Since loops and reductions over shared memory generally require thread synchronization, they should not be used in @unique_shmem_thread expressions.

ClimaComms.unique_shmem_thread — Function

unique_shmem_thread(f, itr)

Functional form of @unique_shmem_thread, equivalent to the expression @unique_shmem_thread itr f().

ClimaComms.shmem_reduce! — Function

shmem_reduce!(op, [dest], itr; [init])

Device-flexible version of reduce(op, itr; [init]), where itr is a shareable iterator that supports setindex!. When dest is also specified, the reduced value is written to it; otherwise, the value is returned at the end of the function.

On GPU devices, shmem_reduce! is parallelized across the threads in each block, so that its runtime that is roughly proportional to log(length(itr)), as long as the block size is no smaller than length(itr). To enable this parallelization, the itr is used as a cache for intermediate data, so the values in it should not be used after this function is called. On CPU devices, shmem_reduce! is equivalent to the standard reduce function, and the itr is not overwritten. The order in which shmem_reduce! evaluates op is device-dependent, so reductions of floating-point values are not always guaranteed to be bitwise reproducible.

ClimaComms.shmem_mapreduce! — Function

shmem_mapreduce!(f, op, [dest], itr; [init])

Device-flexible version of mapreduce(f, op, itr; [init]), where itr is a shareable iterator that supports setindex!. When dest is also specified, the reduced value is written to it; otherwise, the value is returned at the end of the function.

On GPU devices, shmem_mapreduce! is a combination of map!(f, itr, itr) and shmem_reduce!(op, itr; init). On CPU devices, shmem_mapreduce! is equivalent to the standard mapreduce function, and the itr is not overwritten.

ClimaComms.shmem_any! — Function

shmem_any!([f], [dest], itr)

Device-flexible version of any([f], itr), equivalent to calling shmem_reduce! or shmem_mapreduce! with the | operator.

ClimaComms.shmem_all! — Function

shmem_all!([f], [dest], itr)

Device-flexible version of all([f], itr), equivalent to calling shmem_reduce! or shmem_mapreduce! with the & operator.

ClimaComms.shmem_sum! — Function

shmem_sum!([f], [dest], itr; [init])

Device-flexible version of sum([f], itr; [init]), equivalent to calling shmem_reduce! or shmem_mapreduce! with the + operator.

ClimaComms.shmem_prod! — Function

shmem_prod!([f], [dest], itr; [init])

Device-flexible version of prod([f], itr; [init]), equivalent to calling shmem_reduce! or shmem_mapreduce! with the * operator.

ClimaComms.shmem_maximum! — Function

shmem_maximum!([f], [dest], itr; [init])

Device-flexible version of maximum([f], itr; [init]), equivalent to calling shmem_reduce! or shmem_mapreduce! with the max operator.

ClimaComms.shmem_minimum! — Function

shmem_minimum!([f], [dest], itr; [init])

Device-flexible version of minimum([f], itr; [init]), equivalent to calling shmem_reduce! or shmem_mapreduce! with the min operator.

Contexts

ClimaComms.AbstractCommsContext — Type

AbstractCommsContext

The base type for a communications context. Each backend defines a concrete subtype of this.

ClimaComms.SingletonCommsContext — Type

SingletonCommsContext(device=device())

A singleton communications context, used for single-process runs. ClimaComms.AbstractCPUDevice and ClimaComms.CUDADevice device options are currently supported.

ClimaComms.MPICommsContext — Type

MPICommsContext()
MPICommsContext(device)
MPICommsContext(device, comm)

A MPI communications context, used for distributed runs. AbstractCPUDevice and CUDADevice device options are currently supported.

ClimaComms.AbstractGraphContext — Type

AbstractGraphContext

A context for communicating between processes in a graph.

ClimaComms.context — Function

ClimaComms.context(device=device())

Construct a default communication context.

By default, it will try to determine if it is running inside an MPI environment variables are set; if so it will return a MPICommsContext; otherwise it will return a SingletonCommsContext.

Behavior can be overridden by setting the CLIMACOMMS_CONTEXT environment variable to either MPI or SINGLETON.

ClimaComms.graph_context — Function

graph_context(context::AbstractCommsContext,
              sendarray, sendlengths, sendpids,
              recvarray, recvlengths, recvpids)

Construct a communication context for exchanging neighbor data via a graph.

Arguments:

context: the communication context on which to construct the graph context.

sendarray: array containing data to send

sendlengths: list of lengths of data to send to each process ID

sendpids: list of processor IDs to send

recvarray: array to receive data into

recvlengths: list of lengths of data to receive from each process ID

recvpids: list of processor IDs to receive from

This should return an AbstractGraphContext object.

Adapt.adapt_structure — Method

Adapt.adapt_structure(::Type{<:AbstractArray}, context::AbstractCommsContext)

Adapt a given context to a context with a device associated with the given array type.

Example

Adapt.adapt_structure(Array, ClimaComms.context(ClimaComms.CUDADevice())) -> ClimaComms.CPUSingleThreaded()

Note

By default, adapting to Array creates a CPUSingleThreaded device, and there is currently no way to conver to a CPUMultiThreaded device.

Logging

ClimaComms.OnlyRootLogger — Function

OnlyRootLogger()
OnlyRootLogger(ctx::AbstractCommsContext)

Return a logger that silences non-root processes.

If no context is passed, obtain the default context via context.

ClimaComms.MPILogger — Function

MPILogger(context::AbstractCommsContext)
MPILogger(iostream, context)

Add a rank prefix before log messages.

Outputs to stdout if no IOStream is given.

ClimaComms.FileLogger — Function

FileLogger(context, log_dir; log_stdout = true, min_level = Logging.Info)

Log MPI ranks to different files within the log_dir.

The minimum logging level is set using min_level. If log_stdout = true, root process logs will be sent to stdout as well.

Context operations

ClimaComms.init — Function

(pid, nprocs) = init(ctx::AbstractCommsContext)

Perform any necessary initialization for the specified backend. Return a tuple of the processor ID and the number of participating processors.

ClimaComms.mypid — Function

mypid(ctx::AbstractCommsContext)

Return the processor ID.

ClimaComms.iamroot — Function

iamroot(ctx::AbstractCommsContext)

Return true if the calling processor is the root processor.

ClimaComms.nprocs — Function

nprocs(ctx::AbstractCommsContext)

Return the number of participating processors.

ClimaComms.abort — Function

abort(ctx::CC, status::Int) where {CC <: AbstractCommsContext}

Terminate the caller and all participating processors with the specified status.

Collective operations

ClimaComms.barrier — Function

barrier(ctx::CC) where {CC <: AbstractCommsContext}

Perform a global synchronization across all participating processors.

ClimaComms.reduce — Function

reduce(ctx::CC, val, op) where {CC <: AbstractCommsContext}

Perform a reduction across all participating processors, using op as the reduction operator and val as this rank's reduction value. Return the result to the first processor only.

ClimaComms.reduce! — Function

reduce!(ctx::CC, sendbuf, recvbuf, op)
reduce!(ctx::CC, sendrecvbuf, op)

Performs elementwise reduction using the operator op on the buffer sendbuf, storing the result in the recvbuf of the process. If only one sendrecvbuf buffer is provided, then the operation is performed in-place.

ClimaComms.allreduce — Function

allreduce(ctx::CC, sendbuf, op)

Performs elementwise reduction using the operator op on the buffer sendbuf, allocating a new array for the result. sendbuf can also be a scalar, in which case recvbuf will be a value of the same type.

ClimaComms.allreduce! — Function

allreduce!(ctx::CC, sendbuf, recvbuf, op)
allreduce!(ctx::CC, sendrecvbuf, op)

Performs elementwise reduction using the operator op on the buffer sendbuf, storing the result in the recvbuf of all processes in the group. Allreduce! is equivalent to a Reduce! operation followed by a Bcast!, but can lead to better performance. If only one sendrecvbuf buffer is provided, then the operation is performed in-place.

ClimaComms.bcast — Function

bcast(ctx::AbstractCommsContext, object)

Broadcast object from the root process to all other processes. The value of object on non-root processes is ignored.

Graph exchange

ClimaComms.start — Function

start(ctx::AbstractGraphContext)

Initiate graph data exchange.

ClimaComms.progress — Function

progress(ctx::AbstractGraphContext)

Drive communication. Call after start to ensure that communication proceeds asynchronously.

ClimaComms.finish — Function

finish(ctx::AbstractGraphContext)

Complete the communications step begun by start(). After this returns, data received from all neighbors will be available in the stage areas of each neighbor's receive buffer.